Created Character class Test
Created the Character class Test file. As a note: The runAway method can't be evaluated with an assertEquals test due to its RNG for deciding the outcome of if the character escapes. To account for this I made the test simply check to make sure the response isn't null, though it did work with the assertEquals test when the desired outcome was achieved.